requirements.txt
Flask
SQLAlchemy
PyMySQL
cryptography
python-dotenv
Flask-SQLAlchemy
